+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: nroman <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/22 10:07:10 by nroman            #+#    #+#             */
-/*   Updated: 2018/02/26 21:39:02 by nroman           ###   ########.fr       */
+/*   Updated: 2018/02/27 20:17:55 by nroman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strncmp(char *s1, char *s2, unsigned int n)
+int		ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	if (!n)
@@ -25,5 +25,5 @@ int		ft_strncmp(char *s1, char *s2, unsigned int n)
 			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 		i++;
 	}
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
+	return ((int)((unsigned char)s1[i] - (unsigned char)s2[i]));
 }
